Removed the root node lock
This was done by implementing a ref count directly into the pointer. It is safe because the pointer needs to be accessed via a getter in order to stabilize it anyway, so we can sanitize them. Technically this compare and swap setup is prettty similiar to a spin lock, but it is more efficient because it is possible to directly atomically compare and swap in the uncontested case. On my dual core system this is about a ~10% speed increase for inserts with 10 threads contending.